ie8 fix

+++ Epi.DynamicForms.Core/Fields/Select.cs
@@ -127,7 +127,7 @@ namespace MvcDynamicForms.Fields
 
            // LargestChoiseLength = LargestChoiseLength * _ControlFontSize;
 
-            Font stringFont = new Font(ControlFontStyle,(int) _ControlFontSize);
+            Font stringFont = new Font(ControlFontStyle, _ControlFontSize);
 
             SizeF size = new SizeF() ;
             using (Graphics g = Graphics.FromHwnd(IntPtr.Zero))
